django_app/requirements.txt
django
timezonefinder
pillow
mysqlclient
virtualenv
boto3
django-storages
requests
django-ses
sleekxmpp
celery